Add and Impl Hilt as Dependency Injection
Change all modules to be an Android Module instead Java Module, because
the Hilt compiler needs an Android Module.

Create Hilt modules for the modules :app, :domain, :data, :cache, and
:remote. Also apply the DI from VM to DataSource (Remote and Local).

Change the Hilt ViewModel dependence name in the `Dependencies` from lifecycle to viewModel to be more legible.

Add Hilt compiler in the common.gradle as default library, to be
applied for every module